+++ jsf_diplom/src/com/diplom/kozlov/vessel/view/RowBean.java
@@ -19,12 +19,12 @@ public class RowBean {
 
 	@Setter
 	@Getter
-	private Integer loadCapacity;
+	private Double loadCapacity;
 	@Setter
 	@Getter
-	private Integer averageSpeed;
+	private Double averageSpeed;
 	@Setter
 	@Getter
-	private Integer fuelConsumption;
+	private Double fuelConsumption;
 
 }
